requirements.txt
bcrypt==4.3.0
blinker==1.9.0
bottle==0.13.3
certifi==2022.12.7
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.0
cloudscraper==1.2.71
cryptography==44.0.3
dnspython==2.2.1
Flask==3.1.0
Flask-Bcrypt==1.0.1
future==1.0.0
icmplib==3.0.4
idna==3.10
impacket==0.10.0
itsdangerous==2.2.0
Jinja2==3.1.6
ldap3==2.9.1
ldapdomaindump==0.10.0
MarkupSafe==3.0.2
maxminddb==2.7.0
multidict==6.4.3
propcache==0.3.1
proxy_tools==0.1.0
psutil==7.0.0
pyasn1==0.4.8
pycparser==2.22
pycryptodome==3.22.0
pycryptodomex==3.22.0
pyobjc-core==11.0
pyobjc-framework-Cocoa==11.0
pyobjc-framework-Quartz==11.0
pyobjc-framework-Security==11.0
pyobjc-framework-WebKit==11.0
pyOpenSSL==25.0.0
pyparsing==3.2.3
PyRoxy @ git+https://github.com/MatrixTM/PyRoxy.git@ea0f88dbc0573292ba5672124f69e4e7a31b544d
PySocks==1.7.1
pywebview==5.4
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
yarl==1.20.0
